Added fragment detection model to the package. Deleted Fragment_detection folder. Some code still left undocumented in the package, WIP
